chore(optimize-code): Updated router file and created a utils file to add helper functions code and use it from there and removed unwnated code from the components
